+++ data_safe_haven/external/__init__.py
@@ -4,6 +4,7 @@ from .api.graph_api import GraphApi
 from .interface.azure_container_instance import AzureContainerInstance
 from .interface.azure_ipv4_range import AzureIPv4Range
 from .interface.azure_postgresql_database import AzurePostgreSQLDatabase
+from .interface.pulumi_account import PulumiAccount
 
 __all__ = [
     "AzureApi",
@@ -12,4 +13,5 @@ __all__ = [
     "AzureIPv4Range",
     "AzurePostgreSQLDatabase",
     "GraphApi",
+    "PulumiAccount",
 ]
